make check_connection() DRY
